test: backend - user
models test separated
